finished integrating all 5 points required, need to check for bugs on server
