+++ server/mongo.py
@@ -8,3 +8,4 @@ gameInfoCol = db['gameStatus']
 wordListCol = db['wordList']
 dmCol = db['dm']
 imageCol = db['images']
+notiCol = db['notifications']
