+++ src/pages/style.css
@@ -1,9 +1,11 @@
 #Homepage {
     background-color: light;
     margin: 0;
-    height: 100vh;
+    height: 95vh;
     width: 100vw;
-    color: #363537
+    max-width: 100%;
+    color: #363537;
+    text-align: center;
 }
 
 #Login {
@@ -14,7 +16,6 @@
     padding: 10px;
     margin-top: 100px;
     text-align:center;
-
 }
 
 #Logout {
@@ -40,31 +41,61 @@
 }
 
 .homepage_buttons {
+    width: 20vw;
+    height: 40vh;
     /* position: relative; */
     /* width: fit-content; */
-    float:left;
-    margin-left: 100px;
+    /* float:left; */
     /* background-color: lightblue; */
     border: 3px solid gray;
     border-radius: 10px;
     text-align: center;
     padding: 10px;
-
 }
 
 .online_users {
-    float: right;
+    width: 20vw;
+    height: 40vh;
+    /* float: left; */
     /* width: fit-content; */
     /* background-color: lightcoral; */
-    margin-bottom: 10px;
-    margin-right: 100px;
     border: 3px solid gray;
     border-radius: 10px;
-    padding:5px;
+    text-align: left;
+    padding: 10px;
+    overflow-y: scroll;
+}
+
+.notifications {
+    width: 20vw;
+    height: 40vh;
+    /* width: fit-content; */
+    /* background-color: lightcoral; */
+    border: 3px solid gray;
+    border-radius: 10px;
+    padding: 10px;
+    overflow-y: scroll;
+}
+
+.container_divs{
+    margin-left: auto;
+    margin-right: auto;
+    width: 80%;
+    text-align: center;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    flex-direction: row;
+}
+.container > div {
+    flex: 1;
 }
 
 .home_divs{
     text-align: center;
 }
 
-
+.notiList {
+    text-align: center;
+    font-size: 2vh;
+}
